refactored audio connection + moved voice constants to properties file

+++ src/main/java/com/w1sh/medusa/core/listeners/impl/ReadyListener.java
@@ -28,7 +28,7 @@ public class ReadyListener implements EventListener<ReadyEvent> {
                         .last())
                 .doOnNext(ev -> logger.info("All guilds have been received, the client is fully connected"))
                 .flatMap(ev -> event.getClient().getGuilds().count())
-                .doOnNext(guilds -> logger.info("Currently serving {} guilds", guilds.longValue()))
+                .doOnNext(guilds -> logger.info("Currently serving {} guilds", guilds))
                 .then();
     }
 }
